details hotel feature completing , some code refactoring and layout changes

+++ domain-hotel/src/main/java/com/puroblast/domain_hotel/model/Hotel.kt
@@ -1,13 +1,14 @@
 package com.puroblast.domain_hotel.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class Hotel(
     val id: Int = 0,
     val name: String = "",
     @SerializedName("adress")
     val address: String = "",
-    val minimalPrice: Int = 0,
+    val minimalPrice: BigDecimal = BigDecimal(0),
     val priceForIt: String = "",
     val rating: Int = 0,
     val ratingName: String = "",
